fix(galgame): name every block v2 gates behind include=
The v2 cutover sent no include= and no spoiler=, so catalog answered a bare
id+name row for every block and capped tags at spoiler=none. Nothing errored:
the character and staff modals, 会社 logos and descriptions, 声优, rating
histograms and eight of work 3's tags simply rendered empty.

Each face now names its own vocabulary, tracking catalog's apiv2/collect
specs, and the work and character faces ask for spoiler=major so the reader
can raise the ceiling in the browser. GetCharacter and GetStaff read the v2
payload instead of the v1 wire DTOs, which reconnects the two mappers that
only tests were calling; a credit-name detail carries gender, birth and photo
directly, so GetPerson is gone along with the rest of the v1-shaped dead code.

Every entity slice is initialised, because a nil slice marshals to null and
both modals read detail.links.length without a guard.

Goldens are fresh captures from a live v2 face, with new ones for a company
detail and for a work whose companies carry a logo. The face tests assert the
request, not the response — a dropped include token renders an empty page
rather than a red one, so only the request can catch it.

+++ apps/api/internal/galgame/client/catalog_entity.go
@@ -63,9 +63,11 @@ type catalogCreditGroup struct {
 	Credits  []catalogCreditItem `json:"credits"`
 }
 
+// Score is the bucket's value on the source's own scale, not an index:
+// erogamescape publishes deciles, so it is a number and not a count.
 type catalogRatingBucket struct {
-	Score int `json:"score"`
-	Count int `json:"count"`
+	Score float64 `json:"score"`
+	Count int     `json:"count"`
 }
 
 type catalogRating struct {
@@ -105,8 +107,8 @@ type GalgameStaffGroup struct {
 }
 
 type GalgameRatingBucket struct {
-	Score int `json:"score"`
-	Count int `json:"count"`
+	Score float64 `json:"score"`
+	Count int     `json:"count"`
 }
 
 // GalgameRating is one external source's aggregate. Score sits on that source's
